+++ src/features/trainer-board/components/index.js
@@ -1,8 +1,16 @@
-export { Seg7 } from "./Seg7";
-export { LED } from "./LED";
-export { ToggleSW } from "./ToggleSW";
-export { Breadboard } from "./Breadboard";
-export { WireOverlay } from "./WireOverlay";
-export { TrayIC } from "./TrayIC";
-export { DatasheetPopup } from "./DatasheetPopup";
-export { default as Sec } from "./Sec";
+export { default as BoardHeader } from "./BoardHeader";
+export { default as Toolbar } from "./Toolbar";
+export { default as LeftPanel } from "./LeftPanel";
+export { default as CenterPanel } from "./CenterPanel";
+export { default as RightPanel } from "./RightPanel";
+export { default as StatusBar } from "./StatusBar";
+export { default as DragGhost } from "./DragGhost";
+export { default as LoadCircuitModal } from "./LoadCircuitModal";
+export { default as DatasheetPopup } from "./DatasheetPopup";
+export { default as Breadboard } from "./Breadboard";
+export { default as WireOverlay } from "./WireOverlay";
+export { default as TrayIC } from "./TrayIC";
+export { default as ToggleSW } from "./ToggleSW";
+export { default as LED } from "./LED";
+export { default as Seg7 } from "./Seg7";
+export { default as Section } from "./Section";
