+++ src/features/trainer-board/utils/constants.js
@@ -1,0 +1,18 @@
+// Shared display constants used across the trainer board components.
+export const F = "monospace";
+
+export const LEDCOL = { R: "#ff1100", G: "#00ee44", Y: "#ffcc00", B: "#0099ff" };
+
+// Wire color palette — cycled through as the user draws new wires.
+export const WIRE_COLORS = [
+  "#e63946",
+  "#2196f3",
+  "#4caf50",
+  "#ff9800",
+  "#9c27b0",
+  "#00bcd4",
+  "#ffeb3b",
+  "#ff5722",
+  "#f48fb1",
+  "#80cbc4",
+];
