+++ src/features/trainer-board/utils/icCatalog.js
@@ -1,4 +1,7 @@
 // ── IC Catalog ────────────────────────────────────────────────────
+// Physical/visual properties of every chip available in the component
+// tray: pin count, body color, symbol, and a human-readable description
+// shown in tooltips and the datasheet popup.
 export const ICS = {
   7400: {
     name: "Quad 2-in NAND",
@@ -122,129 +125,4 @@ export const ICS = {
   },
 };
 
-// ════════════════════════════════════════════════════════════════════
-// SIMULATION ENGINE
-// Real pin-level netlist + boolean-algebra propagation.
-//
-// Every hole on the breadboard, every external terminal (switches, push
-// buttons, clock, data-bus monitors...) and every IC pin is a *node
-// reference*. Wires and physical breadboard strips union node references
-// together into electrical nodes. Each IC reads the resolved value of its
-// input-pin nodes and drives its output-pin nodes — exactly like a real
-// simulator (Proteus/Multisim/Logisim) resolves a netlist, just scaled
-// down to what a trainer board needs.
-//
-// NOTE ON ACCURACY: gate-level chips (7400/02/04/08/32/86) and the two
-// flip-flop chips (7474/7476) use their real SN74xx datasheet pinouts.
-// The MSI chips (7483/7485/74138/74151/7447/74193/7495) use pinouts that
-// are datasheet-accurate where the author was confident (7447, 74138,
-// 74151 pin numbers are the real ones) and otherwise use a clearly
-// simplified, internally-consistent numbering — good for teaching
-// boolean behavior, but double-check against a real datasheet before
-// wiring a physical board from it.
-// ════════════════════════════════════════════════════════════════════
-
-// ── Gate primitives ────────────────────────────────────────────────
-const G = {
-  and: (a, b) => (a & b) & 1,
-  or: (a, b) => (a | b) & 1,
-  nand: (a, b) => 1 - ((a & b) & 1),
-  nor: (a, b) => 1 - ((a | b) & 1),
-  xor: (a, b) => (a ^ b) & 1,
-  not: (a) => 1 - (a & 1),
-};
-
-// ── IC_LOGIC: pin-accurate behavior tables ─────────────────────────
-// gates: combinational 2-in (or 1-in for inverters) logic cells.
-// flops: edge-triggered storage cells (D or JK).
-// msi:   custom evaluate() for the multi-pin MSI parts.
-export const IC_LOGIC = {
-  7400: {
-    vcc: 14, gnd: 7, gates: [
-      { in: [1, 2], out: 3, fn: G.nand }, { in: [4, 5], out: 6, fn: G.nand },
-      { in: [9, 10], out: 8, fn: G.nand }, { in: [12, 13], out: 11, fn: G.nand },
-    ]
-  },
-  7402: {
-    vcc: 14, gnd: 7, gates: [
-      { in: [2, 3], out: 1, fn: G.nor }, { in: [5, 6], out: 4, fn: G.nor },
-      { in: [8, 9], out: 10, fn: G.nor }, { in: [11, 12], out: 13, fn: G.nor },
-    ]
-  },
-  7404: {
-    vcc: 14, gnd: 7, gates: [
-      { in: [1], out: 2, fn: G.not }, { in: [3], out: 4, fn: G.not },
-      { in: [5], out: 6, fn: G.not }, { in: [9], out: 8, fn: G.not },
-      { in: [11], out: 10, fn: G.not }, { in: [13], out: 12, fn: G.not },
-    ]
-  },
-  7408: {
-    vcc: 14, gnd: 7, gates: [
-      { in: [1, 2], out: 3, fn: G.and }, { in: [4, 5], out: 6, fn: G.and },
-      { in: [9, 10], out: 8, fn: G.and }, { in: [12, 13], out: 11, fn: G.and },
-    ]
-  },
-  7432: {
-    vcc: 14, gnd: 7, gates: [
-      { in: [1, 2], out: 3, fn: G.or }, { in: [4, 5], out: 6, fn: G.or },
-      { in: [9, 10], out: 8, fn: G.or }, { in: [12, 13], out: 11, fn: G.or },
-    ]
-  },
-  7486: {
-    vcc: 14, gnd: 7, gates: [
-      { in: [1, 2], out: 3, fn: G.xor }, { in: [4, 5], out: 6, fn: G.xor },
-      { in: [9, 10], out: 8, fn: G.xor }, { in: [12, 13], out: 11, fn: G.xor },
-    ]
-  },
-  7474: {
-    vcc: 14, gnd: 7, flops: [
-      { kind: "d", d: 2, clk: 3, pr: 4, clr: 1, q: 5, qb: 6 },
-      { kind: "d", d: 12, clk: 11, pr: 10, clr: 13, q: 9, qb: 8 },
-    ]
-  },
-  7476: {
-    vcc: 5, gnd: 13, flops: [
-      { kind: "jk", j: 4, k: 16, clk: 1, pr: 2, clr: 3, q: 14, qb: 15 },
-      { kind: "jk", j: 9, k: 10, clk: 8, pr: 7, clr: 6, q: 11, qb: 12 },
-    ]
-  },
-  // 4-bit full adder. VCC=5 / GND=12 is the real (unusual) 7483 quirk.
-  7483: {
-    vcc: 5, gnd: 12, msi: "adder",
-    a: [1, 2, 3, 4], b: [6, 7, 8, 9], cin: 10, cout: 11, s: [13, 14, 15, 16]
-  },
-  // 4-bit magnitude comparator. Real datasheet pinout.
-  7485: {
-    vcc: 16, gnd: 8, msi: "comparator",
-    a: [1, 2, 3, 4], b: [5, 6, 7, 9], gtIn: 10, ltIn: 11, eqIn: 12,
-    gtOut: 13, ltOut: 14, eqOut: 15
-  },
-  // 3-to-8 decoder. Real datasheet pinout.
-  74138: {
-    vcc: 16, gnd: 8, msi: "decoder",
-    a: [1, 2, 3], e2a: 4, e2b: 5, e1: 6, y: [15, 14, 13, 12, 11, 10, 9, 7]
-  },
-  // 8-to-1 mux. Real datasheet pinout.
-  74151: {
-    vcc: 16, gnd: 8, msi: "mux",
-    d: [4, 3, 2, 1, 15, 14, 13, 12], s: [11, 10, 9], strobe: 7, y: 5, yb: 6
-  },
-  // BCD -> 7-seg driver. Real datasheet pinout.
-  7447: {
-    vcc: 16, gnd: 8, msi: "bcd7seg",
-    bcd: [7, 1, 2, 6], lt: 3, bi: 4, rbi: 5,
-    seg: { a: 13, b: 12, c: 11, d: 10, e: 9, f: 15, g: 14 }
-  },
-  // 4-bit up/down counter. Simplified-but-consistent numbering.
-  74193: {
-    vcc: 16, gnd: 8, msi: "counter",
-    d: [15, 1, 10, 9], up: 5, down: 4, load: 11, clear: 14,
-    q: [3, 2, 6, 7], carry: 12, borrow: 13
-  },
-  // 4-bit serial/parallel shift register. Simplified numbering (author
-  // not fully confident of the real 7495 pin order — verify vs datasheet).
-  7495: {
-    vcc: 14, gnd: 7, msi: "shiftreg",
-    d: [2, 3, 4, 5], serIn: 1, mode: 6, clk1: 9, clk2: 8, q: [10, 11, 12, 13]
-  },
-};
+export default ICS;
