+++ src/features/trainer-board/utils/index.js
@@ -1,5 +1,4 @@
 export * from "./icCatalog";
-export * from "./netlist";
 export * from "./simulationEngine";
 export * from "./breadboardLayout";
-export * from "./pinoutSummary";
+export * from "./constants";
